+++ pkg/skaffold/initializer/init_test.go
@@ -216,6 +216,7 @@ func TestDoInit(t *testing.T) {
 			name: "helm fails",
 			dir:  "testdata/init/helm-deployment",
 			config: initconfig.Config{
+				Force: true,
 				Opts: config.SkaffoldOptions{
 					ConfigurationFile: "skaffold.yaml.out",
 				},
